feat(redis): cache key expire message usecase

+++ spring-cache/spring-redis-event/spring-redis-event/src/main/kotlin/com/toy/springredisevent/user/domain/UserStatusHash.kt
@@ -13,7 +13,7 @@ data class UserStatusHash(
   val statusName: String,
 
   @TimeToLive
-  val ttlSeconds: Int
+  val ttlSeconds: Long
 )
 
 interface UserStatusHashRepository: CrudRepository<UserStatusHash, String>
